move epoch pair to common
